Add files via upload
This task number 3 in my hashtag#oasisinfobyte internship
he task explained is the creation of a temperature converter web application that can convert temperatures between Celsius and Fahrenheit. 
The presenter then proceeds to explain the HTML, CSS, and JavaScript code used to create the temperature converter. The HTML code includes the creation of input fields for the user to enter the temperature to be converted, as well as buttons to initiate the conversion. The CSS code is used to style the web application and make it visually appealing